+++ cvat/apps/iam/permissions.py
@@ -4,28 +4,22 @@
 # SPDX-License-Identifier: MIT
 
 from __future__ import annotations
+
+import operator
 from abc import ABCMeta, abstractmethod
 from collections import namedtuple
-from enum import Enum, auto
-import operator
-from typing import Any, List, Optional, Sequence, Tuple, cast
-
-from attrs import define, field
-from rest_framework.exceptions import ValidationError, PermissionDenied
+from enum import Enum
+from typing import Any, List, Optional, Sequence, cast
 
 import requests
+from attrs import define, field
 from django.conf import settings
 from django.db.models import Q
+from rest_framework.exceptions import ValidationError
 from rest_framework.permissions import BasePermission
 
 from cvat.apps.organizations.models import Membership, Organization
 from cvat.apps.engine.models import Label, Project, Task, Job, Issue
-from cvat.apps.iam.exceptions import LimitsReachedException
-from cvat.apps.limit_manager.core.limits import (CapabilityContext, LimitManager,
-    Limits, OrgCloudStoragesContext, OrgTasksContext, ProjectWebhooksContext,
-    OrgCommonWebhooksContext,
-    TasksInOrgProjectContext, TasksInUserSandboxProjectContext, UserOrgsContext,
-    UserSandboxCloudStoragesContext, UserSandboxTasksContext)
 from cvat.apps.webhooks.models import WebhookTypeChoice
 
 
@@ -1655,11 +1649,7 @@ class LimitPermission(OpenPolicyAgentPermission):
 class PolicyEnforcer(BasePermission):
     # pylint: disable=no-self-use
     def check_permission(self, request, view, obj):
-        # Some permissions are only needed to be checked if the action
-        # is permitted in general. To achieve this, we split checks
-        # into 2 groups, and check one after another.
-        basic_permissions: List[OpenPolicyAgentPermission] = []
-        conditional_permissions: List[OpenPolicyAgentPermission] = []
+        permissions: List[OpenPolicyAgentPermission] = []
 
         # DRF can send OPTIONS request. Internally it will try to get
         # information about serializers for PUT and POST requests (clone
@@ -1668,33 +1658,14 @@ class PolicyEnforcer(BasePermission):
         # the condition below is enough.
         if not self.is_metadata_request(request, view):
             for perm in OpenPolicyAgentPermission.__subclasses__():
-                basic_permissions.extend(perm.create(request, view, obj))
+                permissions.extend(perm.create(request, view, obj))
 
-            conditional_permissions.extend(LimitPermission.create_from_scopes(
-                request, view, obj, basic_permissions
-            ))
-
-        self._iam_context = request.iam_context
-
-        allow = self._check_permissions(basic_permissions)
-        if allow and conditional_permissions:
-            allow = self._check_permissions(conditional_permissions)
-        return allow
-
-    def _check_permissions(self, permissions: List[OpenPolicyAgentPermission]) -> bool:
         allow = True
-        reasons = []
         for perm in permissions:
             result = perm.check_access()
             allow &= result.allow
-            reasons.extend(result.reasons)
 
-        if allow:
-            return True
-        elif reasons:
-            raise LimitsReachedException(reasons, self._iam_context)
-        else:
-            raise PermissionDenied("not authorized")
+        return allow
 
     def has_permission(self, request, view):
         if not view.detail: